+++ codex-rs/exec-server/src/remote_process.rs
@@ -9,8 +9,8 @@ use crate::ExecProcess;
 use crate::ExecProcessEventReceiver;
 use crate::ExecServerError;
 use crate::StartedExecProcess;
+use crate::client::LazyRemoteExecServerClient;
 use crate::client::Session;
-use crate::environment::LazyRemoteExecServerClient;
 use crate::protocol::ExecParams;
 use crate::protocol::ReadResponse;
 use crate::protocol::WriteResponse;
